Worked on gui layout

+++ src/main/java/com/tttsaurus/ingameinfo/common/api/animation/text/ITextAnimDef.java
@@ -4,5 +4,5 @@ import com.tttsaurus.ingameinfo.common.api.render.renderer.AnimTextRenderer;
 
 public interface ITextAnimDef
 {
-    void calcAnim(AnimTextRenderer.CharInfo[] charInfos, float time, float deltaTime);
+    void calcAnim(AnimTextRenderer.CharInfo[] charInfos, double time, double deltaTime);
 }
